Added Music in OurGame
Both songs loop once in the mp3 file. Both have a pickup line in the beginning that should be taken out of the looping in scripts. Let me know if it's easier if I separate the pickup lines and the looping parts of the songs. SFX coming tomorrow (technically today)